+++ src/Hosting/Starcounter.Hosting/TaskSchedulerImpl.cs
@@ -15,27 +15,26 @@ namespace Starcounter.Hosting {
 
         /// <summary>
         /// </summary>
-        public void Run(ITask task, Byte schedId = StarcounterEnvironment.InvalidSchedulerId) {
+        public UInt32 Run(ITask task, Byte schedId = StarcounterEnvironment.InvalidSchedulerId) {
             unsafe {
                 IntPtr hTask = (IntPtr)GCHandle.Alloc(task, GCHandleType.Normal);
 
-                try {
+                UInt32 e = sccorelib.cm2_schedule(
+                    hsched_,
+                    schedId,
+                    sccorelib_ext.TYPE_RUN_TASK,
+                    0,
+                    0,
+                    0,
+                    (ulong)hTask
+                    );
 
-                    var e = sccorelib.cm2_schedule(
-                        hsched_,
-                        schedId,
-                        sccorelib_ext.TYPE_RUN_TASK,
-                        0,
-                        0,
-                        0,
-                        (ulong)hTask
-                        );
-                    if (e != 0) throw ErrorCode.ToException(e);
-                }
-                catch {
+                // If there was an error - we need to cleanup the task.
+                if (0 != e) {
                     ((GCHandle)hTask).Free();
-                    throw;
                 }
+
+                return e;
             }
         }
     }